fix: moved toaster to layout as it does not play nice with the gatsby-focus-wrapper (#391)
