+++ src/pages/invite.tsx
@@ -9,9 +9,9 @@ import Button from "../components/fundamentals/button"
 import LongArrowRightIcon from "../components/fundamentals/icons/long-arrow-right-icon"
 import MedusaIcon from "../components/fundamentals/icons/medusa-icon"
 import MedusaVice from "../components/fundamentals/icons/medusa-vice"
-import LoginLayout from "../components/login-layout"
 import SigninInput from "../components/molecules/input-signin"
 import SEO from "../components/seo"
+import LoginLayout from "../components/templates/login-layout"
 import useNotification from "../hooks/use-notification"
 import { getErrorMessage } from "../utils/error-messages"
